food files moved to components to correct conflict error
